[master] Fix compilation error (#195)
Definition retrieval after writing an optimised module was not working
(unsurprisingly), so the module is now cloned first, which makes
retrieval unnecessary.  That in turn revealed a problem with type
retention in optimised modules, which was fixed by adding an
extern-scoped function that used the type at the time the type was
defined.