[master] Defer abort() when a function cannot be resolved (#197).
When a macro is called, the current module is cloned and finalised.
LLVM then attempts to resolve all symbols in the module when the
compiler requests the address of the function representing the macro
call, even symbols that are not needed by the macro call.  The lazy
function creator previously aborted if it was not able to resolve a
function, but that's too restrictive in this case.  As a stopgap,
return the address of a function that will fail when it is called.
(This should be modified so that the failure function gives the name
of the missing symbol as well.)