Disable escaping from namespaces for inner functions.
It should be invalid to reference a local variable from a different function,
and the check is a work in progress right now.

For the purpose of catching invalid variable references, a set of new 'function
scopes' has been introduced into Context. Every variable will keep track of its
function scope. A variable can be referenced as a value if and only if its
function scope is equal to the current function scope, or its function scope is
the global scope.